+++ packages/react/src/stories/ActionModal/ActionModal.stories.tsx
@@ -7,7 +7,10 @@ import { DESCRIPTION, DO, DONT, TITLE } from './ActionModalInfo';
 import * as StoryTitle from '../StoryTitle';
 import ActionModal, { ActionModalProps } from '../../components/ActionModal';
 import Typography from '../../components/Typography';
-
+/**
+ * To use the library styles you need to import import 'react-sheet-slide/style.css';
+ * as global styles in your project
+ */
 export default {
   title: 'React/Components/ActionModal',
   component: ActionModal,
@@ -38,8 +41,12 @@ const MODAL_FOOTER = (
 
 const VariantComponent = ({
   isAlwaysModal,
+  hasPadding,
+  hasGreyBackground,
 }: {
   isAlwaysModal?: ActionModalProps['isAlwaysModal'];
+  hasGreyBackground?: ActionModalProps['hasGreyBackground'];
+  hasPadding?: ActionModalProps['hasPadding'];
 }) => {
   const [open, setOpen] = useState(false);
 
@@ -51,6 +58,8 @@ const VariantComponent = ({
       <ActionModal
         open={open}
         isAlwaysModal={isAlwaysModal}
+        hasPadding={hasPadding}
+        hasGreyBackground={hasGreyBackground}
         onDismiss={closeModal}
         title={MODAL_TITLE}
         subtitle={MODAL_SUBTITLE}