+++ packages/react/CHANGELOG.md
@@ -3,6 +3,91 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.17.2](https://bitbucket.org/newcrossdev/newcross-ui/compare/@newcross-ui/react@2.17.1...@newcross-ui/react@2.17.2) (2024-11-14)
+
+
+### Bug Fixes
+
+* **CAPP-000:** action modal padding fix ([#130](https://bitbucket.org/newcrossdev/newcross-ui/issues/130)) ([bd27300](https://bitbucket.org/newcrossdev/newcross-ui/commits/bd2730035001351247271ec07130832d09c1c1f5))
+
+
+
+
+
+## [2.17.1](https://bitbucket.org/newcrossdev/newcross-ui/compare/@newcross-ui/react@2.17.0...@newcross-ui/react@2.17.1) (2024-11-14)
+
+
+### Bug Fixes
+
+* **CAPP-000:** action modal padding fix ([#129](https://bitbucket.org/newcrossdev/newcross-ui/issues/129)) ([f20fce0](https://bitbucket.org/newcrossdev/newcross-ui/commits/f20fce03dfe88f5586f729d1c59eb261e97dba75))
+
+
+
+
+
+# [2.17.0](https://bitbucket.org/newcrossdev/newcross-ui/compare/@newcross-ui/react@2.16.0...@newcross-ui/react@2.17.0) (2024-11-14)
+
+
+### Features
+
+* **CAPP-1743:** updating action modal component ([#128](https://bitbucket.org/newcrossdev/newcross-ui/issues/128)) ([3116323](https://bitbucket.org/newcrossdev/newcross-ui/commits/3116323d74f41810e55e2e98f0871a9afed579a6))
+
+
+
+
+
+# [2.16.0](https://bitbucket.org/newcrossdev/newcross-ui/compare/@newcross-ui/react@2.15.2...@newcross-ui/react@2.16.0) (2024-11-12)
+
+
+### Features
+
+* **CAPP-1717:** action modal no footer fix ([#127](https://bitbucket.org/newcrossdev/newcross-ui/issues/127)) ([a29e289](https://bitbucket.org/newcrossdev/newcross-ui/commits/a29e2890ecaaed3d2491ecb5a86294443fa0dd78))
+
+
+
+
+
+## [2.15.2](https://bitbucket.org/newcrossdev/newcross-ui/compare/@newcross-ui/react@2.15.1...@newcross-ui/react@2.15.2) (2024-11-08)
+
+
+### Bug Fixes
+
+* **CAPP-000:** update ActionModal spacings ([#126](https://bitbucket.org/newcrossdev/newcross-ui/issues/126)) ([f330d7c](https://bitbucket.org/newcrossdev/newcross-ui/commits/f330d7c4e854c2da07c25539dd1354c12f22094d))
+
+
+
+
+
+## [2.15.1](https://bitbucket.org/newcrossdev/newcross-ui/compare/@newcross-ui/react@2.15.0...@newcross-ui/react@2.15.1) (2024-11-06)
+
+**Note:** Version bump only for package @newcross-ui/react
+
+
+
+
+
+# [2.15.0](https://bitbucket.org/newcrossdev/newcross-ui/compare/@newcross-ui/react@2.14.0...@newcross-ui/react@2.15.0) (2024-11-05)
+
+
+### Features
+
+* **CAPP-1671:** update font sizes on badges ([#125](https://bitbucket.org/newcrossdev/newcross-ui/issues/125)) ([531bc09](https://bitbucket.org/newcrossdev/newcross-ui/commits/531bc09ba6542e043259f5438723494673f7abdc))
+
+
+
+
+
+# [2.14.0](https://bitbucket.org/newcrossdev/newcross-ui/compare/@newcross-ui/react@2.13.0...@newcross-ui/react@2.14.0) (2024-10-30)
+
+
+### Features
+
+* **ASK-0000:** create new color tokens ([#116](https://bitbucket.org/newcrossdev/newcross-ui/issues/116)) ([162bb92](https://bitbucket.org/newcrossdev/newcross-ui/commits/162bb929296b858c491e4bc1f7b35912bc59c513))
+
+
+
+
+
 # [2.13.0](https://bitbucket.org/newcrossdev/newcross-ui/compare/@newcross-ui/react@2.12.0...@newcross-ui/react@2.13.0) (2024-10-16)